detail

+++ egs/valle-x-multi/scripts/transcripts_from_tsv_md.py
@@ -8,6 +8,7 @@ import re
 import sys
 import time
 from logging import Formatter
+from pathlib import Path
 
 
 def extract_transcripts(directory, target_dir=''):
@@ -24,8 +25,13 @@ def extract_transcripts(directory, target_dir=''):
                     transcript_filename = os.path.join(target_dir, row['SpeakerID'], f"{os.path.splitext(row['UtteranceID'])[0]}_transcript.txt")
 
                     # Write the sentence to a new file in the clips directory
-                    with open(transcript_filename, 'w', encoding='utf-8') as transcript_file:
-                        transcript_file.write(row['Transcription'])
+                    if not Path(transcript_filename).exists():
+                        try:
+                            with open(transcript_filename, 'w', encoding='utf-8') as transcript_file:
+                                transcript_file.write(row['Transcription'])
+                        except Exception as e:
+                            logging.warning(str(e))
+                            pass
 
                     logging.debug(f"Transcript file created: {transcript_filename}")
